+++ cmd/entrypoint.sh
@@ -1,10 +1,44 @@
-#! /usr/bin/env bash
+#!/usr/bin/env bash
+
+# MIT License
+#
+# (C) Copyright [2022] Hewlett Packard Enterprise Development LP
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+
+# run the specified set of tests
 if [[ "$1" == "smoke" ]]; then
     echo "Running smoke tests..."
-    /src/app/smoke_test.py ${@:2} #just pass along the arguments after smoke | functional.
+    # pass along the smoke test arguments
+    /src/app/smoke_test.py ${@:2}
+    RET=$?
 elif [[ "$1" == "functional" ]]; then
     echo "Running functional tests..."
-    /src/app/functional_test.py ${@:2} #just pass along the arguments after smoke | functional.
+    # pass along the functional test arguments
+    /src/app/functional_test.py ${@:2}
+    RET=$?
 else
-    echo "Unsupported test type"
+    echo "ERROR: Unsupported test type"
+    RET=1
 fi
+
+# kill the istio sidecar after the tests run to save wait time
+curl -s -f -X POST http://localhost:15020/quitquitquit
+
+exit ${RET}